Move BioComputeObject plugin to own directory
