+++ client/src/components/Workflow/Invocation/Export/Plugins/BioComputeObject/BioComputeObjectExportPlugin.js
@@ -1,5 +1,5 @@
-import { InvocationExportPlugin, InvocationExportPluginAction } from "../model";
-import SendForm from "./BioComputeObject/SendForm.vue";
+import { InvocationExportPlugin, InvocationExportPluginAction } from "../../model";
+import SendForm from "./SendForm.vue";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { faDatabase } from "@fortawesome/free-solid-svg-icons";
 library.add(faDatabase);
